requirements.txt
asgiref==3.9.1
Django==5.2.5
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing-extensions==4.14.1
whitenoise==6.9.0
pillow
opencv-python
face_recognition
